backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.1
pydantic>=2.7,<3
sqlparse==0.5.5
tzdata==2025.3
# Core Django
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.9.0
python-dotenv==1.0.1

# Caching & Task Queues
celery==5.4.0
redis==5.2.1
django-redis==5.4.0

# LLM & Agent Core
pydantic==2.10.6
openai==1.63.2
anthropic==0.45.2
httpx==0.28.1
aiohttp==3.11.12

# ML and compatibility extraction
torch
torch-geometric
sentence-transformers==5.3.0
piper-tts==1.4.1

# Testing
pytest==9.0.2
pytest-django==4.12.0
